Create prime.f90
A prime number is defined as any integer greater than one which contains no factors other 
than one and itself. If there is no number in the range (1, N) which can divide N with no 
remainder, then N is prime. For example, five is prime number because 1 and 5 are its only 
positive integer factors.
Write a program which can search and print the prime numbers in the range from 2 to 32767. 
Your program must ask for a lower limit and an upper limit to search for all prime numbers 
between such ranges. Finally, the program needs to display the number of prime numbers 
found