Create npv.f90
In finance, the net present value (NPV) is the measurement of the profitability of investment by 
summing the present values of cash inflows and cash outflows over a period of time. Generally, an 
investment with a positive NPV will be a profitable one, while one with a negative NPV will result in 
a net loss.
1) A table that shows the NPV rounded off to the nearest dollar with r = 0%, 10%, 15%, 20%, 25%, 
displayed in percentage with two decimal digits.
2) Find the accurate value of IRR to five decimal places and print the results of 1) and 2) in the similar 
formats shown in the sample below to a screen. A sample of output is shown below where “x” signifies
your calculated numbers and faded numbers indicate the column.